Separate out the Reconciler fellows and document them
