+++ src/temporal-logic-lib/examples/kubernetes_cluster/spec/mod.rs
@@ -6,4 +6,5 @@ pub mod controller;
 pub mod distributed_system;
 pub mod kubernetes_api;
 pub mod network;
+pub mod reconciler;
 pub mod reconcilers;
